Issue #2994506 First pass at correcting the normalizer class hierarchy
